libutil: Move ensureStackSizeAtLeast and savedStackSize to unix specific files
Exposing unix::savedStackSize in a private header (not installed) is intentional
and will be required for making use of in in vfork() based runProgram2.
