+++ src/libutil/unix/current-process-private.hh
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <cstddef>
+
+namespace nix {
+namespace unix {
+
+extern size_t savedStackSize;
+
+} // namespace unix
+} // namespace nix
